feat: env typescript intelisense
