+++ cli/plasmo/src/features/background-service-worker/bgsw-messaging-declaration.ts
@@ -1,30 +1,11 @@
-import { outputFile, outputJson } from "fs-extra"
-import { readFile } from "fs/promises"
-import json5 from "json5"
+import { outputFile } from "fs-extra"
 import { resolve } from "path"
 
 import type { CommonPath } from "~features/extension-devtools/common-path"
+import { addDeclarationConfig } from "~features/helpers/tsconfig"
 
 export const MESSAGING_DECLARATION_FILENAME = `messaging.d.ts`
 const MESSAGING_DECLARATION_FILEPATH = `.plasmo/${MESSAGING_DECLARATION_FILENAME}`
-
-const addMessagingDeclarationConfig = async (commonPath: CommonPath) => {
-  const tsconfigFilePath = resolve(commonPath.projectDirectory, "tsconfig.json")
-
-  const tsconfigFile = await readFile(tsconfigFilePath, "utf8")
-  const tsconfig = json5.parse(tsconfigFile)
-  const includeSet = new Set(tsconfig.include)
-
-  if (includeSet.has(MESSAGING_DECLARATION_FILEPATH)) {
-    return
-  }
-
-  tsconfig.include = [MESSAGING_DECLARATION_FILEPATH, ...includeSet]
-
-  await outputJson(tsconfigFilePath, tsconfig, {
-    spaces: 2
-  })
-}
 
 export const addMessagingDeclaration = (
   commonPath: CommonPath,
@@ -35,7 +16,7 @@ export const addMessagingDeclaration = (
       resolve(commonPath.dotPlasmoDirectory, MESSAGING_DECLARATION_FILENAME),
       declarationCode
     ),
-    addMessagingDeclarationConfig(commonPath)
+    addDeclarationConfig(commonPath, MESSAGING_DECLARATION_FILEPATH)
   ])
 
 export const createDeclarationCode = (messages: string[], ports: string[]) => `
